Added partner definition based on session identifiers.
The new --partner-definition=2 compares session identifiers. These have to given
explicitly in the protocol specification (Cf. the two examples) as

  CLAIM (role, SID, sessionidentifier )

These must be given as early as possible in the spec. Usually this is after
the first receive.

Added explicit static check for SID claims in all roles for SID-partnering.

Added SIDs to Yahalom-Paulson.