BUGFIX: Fixed crash on some include file cases.
Reported by ETH students last year: if you include a file, where the file has an
error in a line with a number higher than the original, the Python code crashes.

This is a *patch* only because the real underlying problem is that error reporting
does not take include commands into account, and does not propagate any
file names.