Added two versions of the NAXOS protocol.
Stopped because I find it hard to understand the required equivalences.
I need to read the paper better first. Now it seems the keys will not
match up at all. I assume they will match up.

It's a bit different as it is missing the combined hash component, and
is not doing any fancy group exponentiation.

Strangely enough --compromise=2 --partner-definition=1 breaks the secrecy claims.
In particular the attack on the responder seems reasonable (not Alice talks to Alice).