Session state inference now appends local state after sends instead of prepends.
The old behaviour meant that any final send of an MH partner could be cut off
and impersonated by the adversary using the local state. This new inference
considers the state to atomically change with the event. Hence, a send is immediately
triggered with the local state change. This is a better approximation.