Added --LKRactorrnsafe switch.
The KCI implementation of the eCK model is also partial: here, we also have that
the RNR query must not have been triggered before. This was not implemented yet,
and is now.

It is possible to make this the default behaviour of the LKRactor switch, which muddles the
LKRactor rule but reduces the number of rules.