Added --force-regular switch. (Hack assistance)
This switch forces the algorithm to also consider a constructed 'regular' session
of the protocols. For the vast majority of protocols, this does not make any difference
other than increasing verification time.

It is useful for DH-like or other hacked protocols where sends and receives with the same
label cannot be unified. This can cause attacks that exploit 'regular executions'
to be missed. This switch would force considering such attacks even though
the protocol hacks themselves might have excluded them.