SCRIPTING ADVERSARY MODELS: Changed transitive closure logic.
The problem with the previous model was that if a claim did not have an attack,
also any weaker models were marked with 'no attack'. This is not necessarily
true. (Of course it holds for the other way around.)

This patch only propagates attacks to other models. Unfortunate side effect:
verification takes more processing time.