Temporary attempt at restoring the attack buffer.
The ancient version (as in the C standard libs) no longer worked under Windows Vista so
this was disabled a while ago.

Now inserted win32-specific code to work around this problem. Sigh.

As a side effect, re-enabled the attack buffer code. This should make
in particular the GUI a bit more responsive in some cases.