Bugfix: manual insertion of (state) compromise events was not working correctly.
In case of manual insertion of compromise events, the compromise type was not properly marked
because the marking code was only triggered when combined with automatic state inference.
Now the marking code is directly connected to the compromise event detection in
the compiler so this should be okay now.

Thanks to Adrian Kyburz for detecting and reporting the bug.