Reintroduced partial first-origination checks for compromise version.
We need to be a bit more careful as properties here (i.e. the adversary
rules) are sensitive to reordering of long-term key reveals, and in
general patterns in which terms are bound from compromise events to not
make other patterns, in which the same terms are bound to a later event,
redundant.