Added option to the Python interface to Scyther to force not using the cache.
The verify() method of the Scyther.Scyther object now has an option
'useCache'. By default, it is True. If set to False, the cache will not
be used.

The scyther-test.py script currently uses this mechanism, which helps
during development, because it simply takes the command-line arguments
and not the file contents into the verify procedure, which implies that
the cache hash does not include the file contents.