BUGFIX: LKRactor rule implementation fixed.
The previous LKRactor rule did not take into account that we are
reasoning with patterns, which can contain variables.
Fixed by a complete reimplementation of LKRactor, that makes an up-front case distinction
on who is the actor of the test run.

This bug was discovered while revisiting the Wimax/IEEE 802.16e pkmv2rsa
fixed protocol version.