Added initial version of cpp conversion script.
Before Scyther directly supported macros, we often relied on external
preprocessors (cpp, or sometimes m4) to perform macro expansion.
In most cases, this is no longer needed.

This script attempts to convert the (cpp-based) input file into an
'.mspdl' file that is parsed internally in Scyther in exactly the same
way as the original using cpp.

The script attempts to check if the internal interpretation in Scyther between
the old and new versions differ, and reports any problems.